requirements.txt
bottle-boilerplate==0.3
